Add 64-bit PMON for Yeloong Platform
Under Yeloong Platform(64-bit PMON):
1. Fix the bug of poweroff command.
2. Fix the bug of loading linux kernel.
3. Fix the bug of going linux kernel.
4. Fix the bug of Update PMON.
5. Fix the bug of unable accessing usb-disk.
6. Add Memtest-86 v3.2 to test almost whole memory under PMON.

NOTE:
  Other Platform can't be compiled at this commitment,this problem
will be solved at the later commitment.

+++ pmon/cmds/newmt/test.h
@@ -125,7 +125,7 @@ void aprint(int y,int x,ulong page);
 void dprint(int y,int x,ulong val,int len, int right);
 void movinv1(int iter, ulong p1, ulong p2);
 void movinvr();
-void movinv32(int iter, ulong p1, ulong lb, ulong mb, int sval, int off);
+void movinv32(int iter, u_int32_t p1, u_int32_t lb, u_int32_t mb, int sval, int off);
 void modtst(int off, int iter, ulong p1, ulong p2);
 void error(ulong* adr, ulong good, ulong bad);
 void ad_err1(ulong *adr1, ulong *adr2, ulong good, ulong bad);
